fix(agents): split origin-guard tests into their own file
render_tests.rs plus the origin-guard tests pushed it over the repo's
600-line max-source-file-length check. No behavior change — same
tests, moved to origin_guard_tests.rs.
